# services/cortex-api/openapi.yaml
openapi: "3.1.0"
info:
  title: cortex-api
  version: "1.0.0"
  description: >
    REST API for the Cortex knowledge system. Cortex holds entities and
    relationships (graph), epistemic state (confidence taxonomy), operational
    intent (todos), episodic memory (session journals), temporal awareness
    (deadlines with urgency), and provenance (evidence URIs). cortex-api is
    the sole access path to cortex.db and todos.db — no direct
    sql(db="cortex") from agents.
servers:
  - url: http://localhost
    description: >
      HTTP target for request paths below. The live service binds a Unix domain
      socket only at /tmp/universal-protocol/cortex-api.sock (see
      docker/compose/cortex-api.yml). Example:
      curl --unix-socket /tmp/universal-protocol/cortex-api.sock http://localhost/health
components:
  schemas:
...
    SessionJournalCreate:
      type: object
      required: [timestamp, agent, summary]
      properties:
        timestamp: { type: string }
        agent: { type: string }
        summary: { type: string }
        domains:
          type: array
          items: { type: string }
          nullable: true
        decisions:
          type: array
          items: { type: string }
          nullable: true
        open_items:
          type: array
          items: { type: string }
          nullable: true
        file_path: { type: string, nullable: true }
    SessionJournalItem:
      type: object
      properties:
        id: { type: integer }
        timestamp: { type: string }
        agent: { type: string }
        summary: { type: string }
        domains:
          type: array
          items: { type: string }
          nullable: true
        decisions:
          type: array
          items: { type: string }
          nullable: true
        open_items:
          type: array
          items: { type: string }
          nullable: true
        file_path: { type: string, nullable: true }
    SessionJournalList:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SessionJournalItem" }
    SessionCloseRequest:
      type: object
      required: [session_id, agent, transcript_md, summary]
      properties:
        session_id:
          type: string
          description: "Pattern: {agent}-YYYY-MM-DD-HHMM"
        agent: { type: string }
        transcript_md:
          type: string
          description: >
            Verbatim dual-layer transcript. Must contain a structural layer
            (## Turn N or ## Session Summary) AND a verbatim layer (### User
            blocks). Hollow structures (no ### User blocks) return 422.
        summary:
          type: string
          description: ">= 20 chars; synthesis of session"
        domains:
          type: array
          items: { type: string }
          nullable: true
        decisions:
          type: array
          items: { type: string }
          nullable: true
        open_items:
          type: array
          items: { type: string }
          nullable: true
        entity_ids:
          type: array
          items: { type: string }
          nullable: true
        prior_session_id:
          type: string
          nullable: true
          description: >
            Previous same-agent session id; creates a `continues` edge from
            this transcript to the prior. Omit only when this is the first
            session in an arc.
        handoff_prompt:
          type: string
          nullable: true
          description: >
            Agent-authored continuation prompt for the next session. Lands
            server-side as a reflective journal entry (kind="handoff",
            register="self") linked to this transcript via
            link_type="handoff_for". Renders verbatim in the next boot
            card's Last Session block. Omission triggers a non-blocking
            post-close `missing_handoff` warning. See
            agent-skills/web-session-close.md Step 2c.
        dry_run:
          type: boolean
          default: false
          description: "Validate-only, no file or DB write"
    SessionCloseResponse:
      type: object
      required: [transcript_entity_id, transcript_path, journal_row_id, session_id]
      properties:
        transcript_entity_id:
          type: string
          description: "e.g. transcript:web-2026-05-04-0844"
        transcript_path: { type: string }
        journal_row_id: { type: integer }
        session_id: { type: string }
        handoff_entry_id:
          type: integer
          nullable: true
          description: >
            Reflective journal entry id when handoff_prompt was provided;
            null otherwise. Use as direct reference to the handoff entry
            for downstream tooling (e.g. arc-view handoffs queries).
        transcript_warnings:
          type: array
          items: { type: string }
          description: "Advisory structural warnings (non-blocking)"
        _warning:
          type: object
          additionalProperties: true
          nullable: true
          description: >
            Audit warnings from pre-close gate (warn mode) and/or post-close
            detectors. Post-close findings under `post_close_findings`
            include `missing_handoff` and `prior_session_id_omitted` kinds.
...
paths:
...
  /session-journals:
    get:
      operationId: listSessionJournals
      summary: List recent session journals
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 3 }
      responses:
        "200":
          description: Session journal list
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionJournalList" }
    post:
      operationId: createSessionJournal
      summary: Create a session journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SessionJournalCreate" }
      responses:
        "201":
          description: Created session journal
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionJournalItem" }
  /session-journals/close:
    post:
      operationId: closeSession
      summary: Atomic session close (transcript + entity + journal + continues edge + optional handoff)
      description: >
        Single-transaction session close. Writes transcript file, creates
        `transcript:{session_id}` entity, inserts the session journal row,
        creates a `continues` edge to `prior_session_id` if provided, and
        optionally inserts a reflective journal handoff entry linked to the
        transcript. Rejects stub transcripts and dual-layer-hollow transcripts
        with 422. Post-close detectors (`missing_handoff`,
        `prior_session_id_omitted`) attach warnings to `_warning` but do not
        block. See agent-skills/web-session-close.md and
        agent-skills/session-close-audit.md.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SessionCloseRequest" }
      responses:
        "201":
          description: Session closed atomically
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionCloseResponse" }
        "422":
          description: >
            Transcript validation failed. Reason codes:
            transcript.hollow, transcript.missing_structure,
            summary.too_short, session_id.invalid.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
                  reason: { type: string }
...